+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 -   `isOpen` parameter for state `onInit`.
 -   Responsive `data-open` for each size: `data-md-open` ...
 -   `dropy.config.set` object for set dropy config.
+-   `opened`,`closed`,`inited` states.
 
 ## [v1.2.0](https://github.com/amirHossein5/dropy/compare/v1.1.0...v1.2.0)